third_party/WebKit/LayoutTests/shadow-dom/imperative-api.html

<!DOCTYPE html>
<!--
Tests for Imperative Shadow DOM Distribution API.
See https://crbug.com/869308
-->
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="host">
</div>
<div id="host1">
</div>

<script>
const host = document.querySelector('#host');
const host1 = document.querySelector('#host1');

test(() => {
  assert_not_equals(host.attachShadow({ mode: 'open', slotting: 'manual' }),
                    null, 'slotting manual should work');
  assert_not_equals(host1.attachShadow({ mode: 'open', slotting: 'auto' }),
                    null, 'slotting auto should work');
  assert_throws(new TypeError(), () => {
    host.attachShadow({ mode: 'open', slotting: 'exceptional' })},
                'others should throw exception');
}, 'attachShadow can take slotting parameter');
</script>
